Hyunji

+++ src/main/java/com/itd5/homeReviewSite/model/keyword.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Entity
 public class keyword {
     @Id
-    private int review_id;
+    private Long review_id;
     private double noise;
     private double smell;
     private double safety;
